fastAPI/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.6.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
requests==2.31.0
aiofiles==23.2.1
httpx==0.26.0
